Move logic for external vocabularies into its class.
