+++ src/main/java/cz/cvut/kbss/termit/service/business/VocabularyService.java
@@ -302,160 +302,6 @@ public class VocabularyService
         eventPublisher.publishEvent(new VocabularyCreatedEvent(this, imported.getUri()));
         return imported;
     }
-
-    
-    private static final String LIST_AVAILABLE_VOCABULARIES_QUERY = "import/listAvailableVocabularies.rq";
-    private static final String EXPORT_FULL_VOCABULARY_QUERY = "import/exportFullVocabulary.rq";
-
-/**
- * Sends a SPARQL query to fetch list of available vocabularies.
- * 
- * @return list of available vocabulary information or null if connection failed
- */
-    public List<RdfsResource> getAvailableVocabularies() {
-
-        List<RdfsResource> response = new ArrayList<>();
-        
-        final Configuration config = context.getBean(Configuration.class);
-
-        String sparqlEndpoint = config.getExternal().getResource();
-        String sparqlQuery = Utils.loadQuery(LIST_AVAILABLE_VOCABULARIES_QUERY);
-
-        SPARQLRepository sparqlRepo = new SPARQLRepository(sparqlEndpoint);
-        sparqlRepo.init();
-        try (RepositoryConnection conn = sparqlRepo.getConnection()) {
-            TupleQuery query = conn.prepareTupleQuery(sparqlQuery);
-
-            try (TupleQueryResult result = query.evaluate()) {
-                while (result.hasNext()) {
-                    BindingSet line = result.next();
-                    if (!line.hasBinding("slovnik")) {
-                        LOG.error("Error: no slovnik binding in: {}", line.toString());
-                        continue;
-                    }
-                    URI uri = new URI(line.getBinding("slovnik").getValue().stringValue());
-                    HashMap<String, String> labels = new HashMap<>();
-                    
-                    // add cs label if available
-                    if (line.hasBinding("nazev_slovniku_cs")) {
-                        labels.put("cs", line.getBinding("nazev_slovniku_cs").getValue().stringValue());
-                    } else {
-                        labels.put("cs", uri.toString());
-                    }
-                    // add en label if available
-                    if (line.hasBinding("nazev_slovniku_en")) {
-                        labels.put("en", line.getBinding("nazev_slovniku_en").getValue().stringValue());
-                    } else {
-                        labels.put("en", uri.toString());
-                    }
-                    MultilingualString label = new MultilingualString(labels);
-                    response.add(new RdfsResource(uri, label, new MultilingualString(), ""));
-                }
-            }
-        } catch (Exception e) {
-            LOG.error(e.getMessage());
-        }
-        return response;
-    }
-
-    /**
-     * Imports multiple vocabularies from external source.
-     *
-     * @param vocabularyIris List of
-     * @return first imported Vocabulary
-     * @throws URISyntaxException
-     * @throws QueryEvaluationException
-     * @throws RepositoryException
-     */
-    @Transactional
-    @PreAuthorize("@vocabularyAuthorizationService.canCreate()")
-    public Vocabulary importFromExternalUris(List<String> vocabularyIris) throws URISyntaxException, QueryEvaluationException, RepositoryException {
-        Vocabulary firstImportedVocabulary = null;
-        
-        for (String vocabularyIri : vocabularyIris) {
-            InputStream newVocabulary = downloadExternalVocabulary(vocabularyIri);
-            if (newVocabulary != null) {
-                URI uri = new URI(vocabularyIri);
-                Vocabulary vocabulary = repositoryService.importVocabulary(uri, RDFFormat.TURTLE.getDefaultMIMEType(), newVocabulary);
-
-                // add types
-                vocabulary.addType(cz.cvut.kbss.termit.util.Vocabulary.s_c_pouze_pro_cteni);
-                vocabulary.addType(cz.cvut.kbss.termit.util.Vocabulary.s_c_externi);
-
-                final AccessControlList acl = aclService.createFor(vocabulary);
-                vocabulary.setAcl(acl.getUri());
-
-                eventPublisher.publishEvent(new VocabularyCreatedEvent(this, vocabulary.getUri()));
-                LOG.debug("Vocabulary {} import was successful.", vocabularyIri);
-                if (firstImportedVocabulary == null) {
-                    firstImportedVocabulary = vocabulary;
-                }
-            }
-        }
-        
-        return firstImportedVocabulary;
-    }
-
-    @Transactional
-    private InputStream downloadExternalVocabulary(String vocabularyIri) throws QueryEvaluationException, RepositoryException{
-
-        List<RdfsResource> response = new ArrayList<>();
-        final Configuration config = context.getBean(Configuration.class);
-
-        String sparqlEndpoint = config.getExternal().getResource();;
-        String sparqlQuery = String.format(Utils.loadQuery(EXPORT_FULL_VOCABULARY_QUERY), vocabularyIri);
-
-        SPARQLRepository sparqlRepo = new SPARQLRepository(sparqlEndpoint);
-        sparqlRepo.init();
-        RepositoryConnection conn = sparqlRepo.getConnection();
-        GraphQuery graphQuery = conn.prepareGraphQuery(sparqlQuery);
-
-        GraphQueryResult result = graphQuery.evaluate();
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        Rio.write(result, outputStream, RDFFormat.TURTLE);
-        InputStream vocabularyFile = new ByteArrayInputStream(outputStream.toByteArray());
-        return vocabularyFile;
-
-    }
-    
-    @Scheduled(cron = "0 0 0 * * *") // Every day at midnight
-    public void reloadVocabularies() throws URISyntaxException {
-        System.out.println("Reloading vocabularies at " + Instant.now());
-        try {
-            // Create domain user
-            cz.cvut.kbss.termit.model.User domainUser = new cz.cvut.kbss.termit.model.User();
-            domainUser.setUsername("system");
-
-            // Create TermItUserDetails (or use a factory if available)
-            TermItUserDetails userDetails = new TermItUserDetails(domainUser.toUserAccount(), List.of());
-
-            // Set auth
-            UsernamePasswordAuthenticationToken auth =
-                new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-
-            SecurityContextHolder.getContext().setAuthentication(auth);
-
-            reloadAllExternal();
-        } catch (Exception e){
-            throw e;
-        } finally {
-            // Clear security context afterwards
-            SecurityContextHolder.clearContext();
-        }
-//        reloadAllExternal();
-    }
-    
-    
-    public void reloadAllExternal() throws URISyntaxException{
-        List<String> externalVocabularies = findAll().stream()
-                .filter((t) -> t.getTypes().contains(cz.cvut.kbss.termit.util.Vocabulary.s_c_externi))
-                .map((t) -> t.getUri().toString())
-                .toList();
-//        Dont have the right to import vocabularies.
-//        importFromExternalUris(externalVocabularies);
-        
-    }
-    
     
     /**
      * Imports a vocabulary from the specified file.
